Document-Search-Engine-main/requirements.txt
# LangChain Core
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
docx2txt>=0.8

# Vector Store
docarray>=0.40.0

# Embeddings
sentence-transformers>=2.2.2

# API & Environment
python-dotenv>=1.0.0
groq>=0.4.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
Werkzeug>=3.0.0
